MAINT: Separate components dir between maps and schedule
