+++ src/components/modules/Map/PlaceMarker.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useTheme } from '@mui/material/styles'
 import { Marker } from '@react-google-maps/api'
 
-import { SpotDTO } from './SpotCard'
+import { SpotDTO } from '../SpotCard'
 
 type Props = {
   name: string
