+++ src/components/modules/Map/SearchBox.tsx
@@ -16,8 +16,8 @@ import { faClose, faSearch } from '@fortawesome/free-solid-svg-icons'
 import InfiniteScroll from 'react-infinite-scroll-component'
 
 import { useGetSpotsWithMatchingNameLazyQuery } from 'generated/graphql'
-import SpotsList from './SpotsList'
-import type { SpotDTO } from './SpotCard'
+import SpotsList from '../SpotsList'
+import type { SpotDTO } from '../SpotCard'
 import { useToggle } from 'react-use'
 import { usePlaces } from 'hooks/googlemaps/usePlaces'
 import { useGoogleMap } from '@react-google-maps/api'
