+++ src/components/modules/Schedule/AddEventCard.tsx
