+++ src/components/modules/Schedule/HomeEventCard.tsx
