+++ src/components/modules/Schedule/SpotEventCard.tsx
