+++ src/components/modules/Schedule/SpotLabel.tsx
